Added ops triggers

+++ lib/cart_model.dart
@@ -2,8 +2,8 @@ class Cart {
   late final int? id;
   final String? productId;
   final String? productName;
-  final int? initialPrice;
-  final int? productPrice;
+  final double? initialPrice;
+  final double? productPrice;
   final int? quantity;
   final String? unitTag;
   final String? image;
